feat: add BusinessLogicException and implement all /lists endpoints

+++ src/main/java/hu/unideb/cartshare/service/ListService.java
@@ -5,13 +5,14 @@ import java.util.UUID;
 
 import org.springframework.stereotype.Service;
 
+import hu.unideb.cartshare.exception.BusinessLogicException;
+import hu.unideb.cartshare.mapper.ListMapper;
 import hu.unideb.cartshare.model.dto.request.ListRequestDto;
 import hu.unideb.cartshare.model.dto.response.ListResponseDto;
-import hu.unideb.cartshare.mapper.ListMapper;
 import hu.unideb.cartshare.model.entity.UserListMembership;
 import hu.unideb.cartshare.model.enums.UserListRole;
-import hu.unideb.cartshare.repository.UserListMembershipRepository;
 import hu.unideb.cartshare.repository.ListRepository;
+import hu.unideb.cartshare.repository.UserListMembershipRepository;
 import lombok.RequiredArgsConstructor;
 
 @Service
@@ -34,19 +35,71 @@ public class ListService {
 
     public ListResponseDto create(ListRequestDto dto) {
         hu.unideb.cartshare.model.entity.List list = mapper.toEntity(dto);
+
         UserListMembership membership = new UserListMembership();
         membership.setList(list);
         membership.setUser(userService.findById(USER_ID));
+        membership.setRole(UserListRole.OWNER);
+
         userListMembershipRepository.save(membership);
         repository.save(list);
+
         return mapper.toDto(list);
     }
 
+    public ListResponseDto joinById(UUID id) {
+        hu.unideb.cartshare.model.entity.List foundList = findById(id);
+
+        if (userListMembershipRepository.existsByListAndUser_IdAndRoleIs(foundList, USER_ID, UserListRole.OWNER)) {
+            throw new BusinessLogicException("Ennek a listának már tulajdonosa vagy!");
+        }
+        if (userListMembershipRepository.existsByListAndUser_IdAndRoleIs(foundList, USER_ID, UserListRole.MEMBER)) {
+            throw new BusinessLogicException("Ennek a listának már tagja vagy!");
+        }
+
+        UserListMembership membership = new UserListMembership();
+        membership.setList(foundList);
+        membership.setUser(userService.findById(USER_ID));
+        membership.setRole(UserListRole.MEMBER);
+        // TODO: joinedAt ?
+        userListMembershipRepository.save(membership);
+
+        return mapper.toDto(foundList);
+    }
+
+    public ListResponseDto update(UUID id, ListRequestDto dto) {
+        hu.unideb.cartshare.model.entity.List foundList = findById(id);
+
+        foundList.setName(dto.getName());
+
+        repository.save(foundList);
+
+        return mapper.toDto(foundList);
+    }
+
+    public ListResponseDto leaveById(UUID id) {
+        hu.unideb.cartshare.model.entity.List foundList = findById(id);
+
+        if (!userListMembershipRepository.existsByListAndUser_IdAndRoleIs(foundList, USER_ID, UserListRole.MEMBER)) {
+            throw new BusinessLogicException("Ennek a listának nem vagy a tagja!");
+        }
+
+        userListMembershipRepository.deleteByListAndUser_Id(foundList, USER_ID);
+
+        return mapper.toDto(foundList);
+    }
+
     public void delete(UUID id) {
+        hu.unideb.cartshare.model.entity.List foundList = findById(id);
+
+        if (!userListMembershipRepository.existsByListAndUser_IdAndRoleIs(foundList, USER_ID, UserListRole.OWNER)) {
+            throw new BusinessLogicException("Ennek a listának nem vagy a tulajdonosa!");
+        }
+
         repository.deleteById(id);
     }
 
-    public ListResponseDto joinById(UUID id) {
-        return null;
+    private hu.unideb.cartshare.model.entity.List findById(UUID id) {
+        return repository.findById(id).orElseThrow(() -> new BusinessLogicException("Ez a lista nem létezik!"));
     }
 }
